LU-13566 socklnd: fix local interface binding
When a node is configured with multiple interfaces in
Multi-Rail config, socklnd was not utilizing the local interface
requested by LNet. In essence LNet was using all the NIDs in round
robin, however the socklnd module was not binding to the correct
interface. Traffic was thus sent on a subset of the interfaces.

The reason is that the route interface number was not being set.
In most cases lnet_connect() is called to create a socket. The
socket is bound to the interface provided and then
ksocknal_create_conn() is called to create the socklnd connection.
ksocknal_create_conn() calls ksocknal_associate_route_conn_locked()
at which point the route's local interface is assigned. However,
this is already too late as the socket has already been created
and bound to a local interface.

Therefore, it's important to assign the route's interface before
calling lnet_connect() to ensure socket is bound to correct local
interface.

To address this issue, the route's interface myipaddr is initialized
to the NI's interface ipaddr when it's added to the peer_ni.

Note: We're strictly assuming that there is one interface for each
NI. This is because tcp bonding will be removed from the socklnd as
it has been deprecated by LNet mutli-rail.

The patch reworked during porting, index was changed to ipaddr,
since it is used at current version of lnet_connect.

Signed-off-by: Amir Shehata <ashehata@whamcloud.com>
Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: Ib336f213f19a0eac93a69367b48417df0ea9b581
Reviewed-on: https://es-gerrit.dev.cray.com/157713
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Reviewed-by: Chris Horn <hornc@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>