LUS-9099 lnet: Use lr_hops for avoid_asym_router_failure
In order for the asymmetric route failure avoidance feature to work
properly it needs to know what the hop count of a route should be.
This information is defined by the lr_hops field of the lnet_route.
The lr_single_hop is what discovery was able to determine the hop
count actually is (single or multi) based on the last ping reply.
If a remote interface on a router goes missing, the route may be
classified as multi-hop by discovery, but it should be considered
single-hop for the purposes of avoiding asymmetric route failure.

Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: I9c255f9a2175d964661850277808dae96ff7735c
Reviewed-on: https://es-gerrit.dev.cray.com/157564
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>
Reviewed-by: Shaun Tancheff <stancheff@cray.com>